ignore dist template

+++ api/App/Data.php
@@ -18,22 +18,24 @@ class Data extends Controller
     $pagePath = DOC_ROOT.Config::get('storage').'/pages/'.$page.'.html';
     @mkdir(DOC_ROOT.Config::get('storage').'/pages', 0777, true);
 
+    if (file_exists(dirname(__FILE__).'/template.dist.php'))
+    {
+      $template_vars = array(
+        'title' => $_POST['title'],
+        'keywords' => $_POST['keywords'],
+        'description' => $_POST['description'],
+        'html' => $_POST['html']
+      );
 
-    $template_vars = array(
-      'title' => $_POST['title'],
-      'keywords' => $_POST['keywords'],
-      'description' => $_POST['description'],
-      'html' => $_POST['html']
-    );
+      extract($template_vars);
 
-    extract($template_vars);
+      ob_start();
+      include dirname(__FILE__).'/template.dist.php';
+      $html = ob_get_contents();
+      ob_end_clean();
 
-    ob_start();
-    include dirname(__FILE__).'/template.dist.php';
-    $html = ob_get_contents();
-    ob_end_clean();
-
-    file_put_contents($pagePath, $html);
+      file_put_contents($pagePath, $html);
+    }
 
     return $this->json(['success' => true, 'page' => $page]);
   }
